Move database related functions to their own module.
